pyproject.toml
[project]
name = "agentic-cert-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "asttokens==3.0.0",
    "attrs==25.4.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "beautifulsoup4==4.14.2",
    "brotli==1.2.0",
    "build==1.3.0",
    "cachetools==6.2.1",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "chromadb==1.3.7",
    "click==8.3.0",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "cryptography==46.0.3",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "distro==1.9.0",
    "durationpy==0.10",
    "emoji==2.15.0",
    "executing==2.2.1",
    "fastapi==0.126.0",
    "ffmpy==1.0.0",
    "filelock==3.20.0",
    "filetype==1.2.0",
    "flatbuffers==25.9.23",
    "frozenlist==1.8.0",
    "fsspec==2025.9.0",
    "google-auth==2.42.0",
    "googleapis-common-protos==1.71.0",
    "gradio==6.2.0",
    "gradio-client==2.0.2",
    "greenlet==3.3.0",
    "groovy==0.1.2",
    "groq==0.33.0",
    "grpcio==1.76.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "html5lib==1.1",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.36.0",
    "humanfriendly==10.0",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "ipykernel==7.0.1",
    "ipython==9.6.0",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets==8.1.8",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.11.1",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.9.1",
    "jupyterlab-widgets==3.0.16",
    "langchain==1.2.0",
    "langchain-chroma==1.1.0",
    "langchain-classic==1.0.0",
    "langchain-community==0.4.1",
    "langchain-core==1.2.4",
    "langchain-groq==1.1.1",
    "langchain-huggingface==1.2.0",
    "langchain-text-splitters==1.1.0",
    "langchain-unstructured==1.0.0",
    "langdetect==1.0.9",
    "langgraph==1.0.5",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.5",
    "langgraph-sdk==0.3.1",
    "langsmith==0.5.0",
    "libmagic==1.0",
    "lxml==6.0.2",
    "markdown==3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.2",
    "matplotlib-inline==0.2.1",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mouseinfo==0.1.3",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.6.1",
    "nltk==3.9.2",
    "numpy==2.3.4",
    "oauthlib==3.3.1",
    "olefile==0.47",
    "onnxruntime==1.19.2",
    "openai==2.6.1",
    "opentelemetry-api==1.38.0",
    "opentelemetry-exporter-otlp-proto-common==1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.38.0",
    "opentelemetry-proto==1.38.0",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-semantic-conventions==0.59b0",
    "orjson==3.11.4",
    "ormsgpack==1.12.1",
    "outcome==1.3.0.post0",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "parso==0.8.5",
    "pillow==12.0.0",
    "platformdirs==4.5.0",
    "posthog==5.4.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "protobuf==6.33.0",
    "psutil==7.1.2",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pyautogui==0.9.54",
    "pybase64==1.4.2",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.12.0",
    "pydub==0.25.1",
    "pygetwindow==0.0.9",
    "pygments==2.19.2",
    "pymsgbox==2.0.1",
    "pypdf==6.5.0",
    "pyperclip==1.11.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pyreadline3==3.5.4",
    "pyrect==0.2.0",
    "pyscreeze==1.0.1",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-iso639==2025.11.16",
    "python-magic==0.4.27",
    "python-multipart==0.0.21",
    "python-oxmsg==0.0.2",
    "pytweening==1.2.0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "rapidfuzz==3.14.3",
    "referencing==0.37.0",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rpds-py==0.28.0",
    "rsa==4.9.1",
    "safehttpx==0.1.7",
    "safetensors==0.7.0",
    "schedule==1.2.2",
    "scikit-learn==1.8.0",
    "scipy==1.16.3",
    "selenium==4.38.0",
    "semantic-version==2.10.0",
    "sentence-transformers==5.2.0",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.8",
    "sqlalchemy==2.0.45",
    "stack-data==0.6.3",
    "starlette==0.50.0",
    "sympy==1.14.0",
    "tavily==1.1.0",
    "tavily-python==0.7.17",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tokenizers==0.22.1",
    "tomlkit==0.13.3",
    "torch==2.9.1",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.57.3",
    "trio==0.31.0",
    "trio-websocket==0.12.2",
    "typer==0.20.0",
    "typer-slim==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "unstructured==0.18.21",
    "unstructured-client==0.42.6",
    "urllib3==2.6.2",
    "uuid-utils==0.12.0",
    "uvicorn==0.38.0",
    "watchfiles==1.1.1",
    "wcwidth==0.2.14",
    "webencodings==0.5.1",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "widgetsnbextension==4.0.15",
    "wrapt==2.0.1",
    "wsproto==1.2.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
